+++ src/layer/middleware/cwa-auth.ts
@@ -1,5 +1,6 @@
 import { defineNuxtRouteMiddleware, navigateTo, useNuxtApp } from 'nuxt/app'
 import type { RouteLocationNormalized } from 'vue-router'
+import { useProcess } from '#cwa/composables/process'
 
 export default defineNuxtRouteMiddleware(async (to: RouteLocationNormalized) => {
   const { $cwa } = useNuxtApp()
@@ -7,6 +8,10 @@ export default defineNuxtRouteMiddleware(async (to: RouteLocationNormalized) => 
   await $cwa.auth.init()
 
   if ($cwa.auth.signedIn.value) {
+    return
+  }
+
+  if (useProcess().isServer) {
     return
   }